add optional reputation recovery suppression
Reputation entries can have an optional DecayAfter value set which is a
timestamp indicating the period of time after which the reputation
should begin to decay. This overrides the default behavior, where the
reputation will begin to decay immediately.

Also adds support for specification of a decay after value via a
recovery suppression field in the violation POST body, which informs
iprepd that it should adjust the DecayAfter value to this point in the
future (in seconds).